+++ Certificates/ca.pem
@@ -1,0 +1,22 @@
+-----BEGIN CERTIFICATE-----
+MIIDujCCAqKgAwIBAgIUes2Ijqwv4MOtimyl/wfYOcizI/kwDQYJKoZIhvcNAQEL
+BQAwdTELMAkGA1UEBhMCVVMxFjAUBgNVBAgTDVNhbiBGcmFuY2lzY28xCzAJBgNV
+BAcTAkNBMRgwFgYDVQQKEw9NeSBDb21wYW55IE5hbWUxEzARBgNVBAsTCk9yZyBV
+bml0IDIxEjAQBgNVBAMTCU15IG93biBDQTAeFw0yMjA1MTgwMDA4MDBaFw0yNzA1
+MTcwMDA4MDBaMHUxCzAJBgNVBAYTAlVTMRYwFAYDVQQIEw1TYW4gRnJhbmNpc2Nv
+MQswCQYDVQQHEwJDQTEYMBYGA1UEChMPTXkgQ29tcGFueSBOYW1lMRMwEQYDVQQL
+EwpPcmcgVW5pdCAyMRIwEAYDVQQDEwlNeSBvd24gQ0EwggEiMA0GCSqGSIb3DQEB
+AQUAA4IBDwAwggEKAoIBAQDYiN2/+UObc5eD51a9l9pMdGZfw9/PG+E7Frc0h1Xf
+mm9h4vd++qdcqlQ/x+5MP1SqY0jwoedi1dYmoIRsz56ezvoZSS5IFpdSPCzSnneL
+Dh4Cphax/bZ8ZjglHKjmzXZNGdoGzD9uN6Wh1CbGYd4abKdMWfWcaq0D5N+vHSYt
+zRotKZ8VOQOJwdtJsXbQ7teR4GqCdS9MhSsg5u+rYR6t6uvS6Ba4p5qlQmv81FZM
+/gIk5wfM05i9BS8y7AGJLym6/VobKhk3hpMevaGHvhanEDPjYHcQ7SamW7fzuGtu
+22oUNfZQSmPApTMqJfhbUj86Is0C+J5Sz5JKuPzcX+9BAgMBAAGjQjBAMA4GA1Ud
+DwEB/wQEAwIBBjAPBgNVHRMBAf8EBTADAQH/MB0GA1UdDgQWBBRJsWpQbxqhmIlR
+opnsbQ/oOg7bdjANBgkqhkiG9w0BAQsFAAOCAQEAsP986LMb0Y5NvCrLZMHm3B9K
+kPvIQEGlrTFcuzymFU8eEFHq9FWSPMEevHOa348ogt5f7t/SgEPcVFW4rr/tPZ03
+fziMSDWCQcD7ne0mH7z7Az9LHvUCNDFsVszqJTKAPb6CE0FSMVL3Nwa3j7QoJduG
+O9iYIoHolpPqSbrdB9dey5SXrx0ooYAN3sCExt53D+p0hsP6WnsLE6gKHeRWO4cK
+8dzwl2QtiP8zcnQrxT1370UdJqme3K3OZZbkH/k0NuJGPnK+UDd/INoSV/913OEv
+vTXAGMsjiEp/payoXBqjAjBBh4vAc5ykh0njkMKHzdocNopax3w7QWeAciOxuA==
+-----END CERTIFICATE-----
