+++ Certificates/client-key.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIK0ac4gROS3cxxGHkbs5lYoWniaIUzkKupg5UNSfyyTboAoGCCqGSM49
+AwEHoUQDQgAEqhHzHkftDwlLubTepIVb23bL0IfOPRhD7Fl6Ei6buNO6VE3yHqpb
+GxsOyvT1k2IC4qmH5qO13lZFJeaweAU6pg==
+-----END EC PRIVATE KEY-----
