+++ Certificates/client.pem
@@ -1,0 +1,19 @@
+-----BEGIN CERTIFICATE-----
+MIIDGzCCAgOgAwIBAgIURqCxvDARE3PUezUdiPszOzPvKskwDQYJKoZIhvcNAQEL
+BQAwdTELMAkGA1UEBhMCVVMxFjAUBgNVBAgTDVNhbiBGcmFuY2lzY28xCzAJBgNV
+BAcTAkNBMRgwFgYDVQQKEw9NeSBDb21wYW55IE5hbWUxEzARBgNVBAsTCk9yZyBV
+bml0IDIxEjAQBgNVBAMTCU15IG93biBDQTAeFw0yMjA1MTgwMDEyMDBaFw0yNzA1
+MTcwMDEyMDBaMEMxCzAJBgNVBAYTAlVTMQswCQYDVQQIEwJDQTEWMBQGA1UEBxMN
+U2FuIEZyYW5jaXNjbzEPMA0GA1UEAxMGY2xpZW50MFkwEwYHKoZIzj0CAQYIKoZI
+zj0DAQcDQgAEqhHzHkftDwlLubTepIVb23bL0IfOPRhD7Fl6Ei6buNO6VE3yHqpb
+GxsOyvT1k2IC4qmH5qO13lZFJeaweAU6pqOBnzCBnDAOBgNVHQ8BAf8EBAMCBaAw
+EwYDVR0lBAwwCgYIKwYBBQUHAwIwDAYDVR0TAQH/BAIwADAdBgNVHQ4EFgQULulR
+sIP07l10oqtDXWR+kSXB9HgwHwYDVR0jBBgwFoAUSbFqUG8aoZiJUaKZ7G0P6DoO
+23YwJwYDVR0RBCAwHoILZXhhbXBsZS5uZXSCD3d3dy5leGFtcGxlLm5ldDANBgkq
+hkiG9w0BAQsFAAOCAQEAdZ1IgRGxRfGzqZMY9iMrYaPcUtMe0spz0Wf68p+QF+og
+Q2yixsKk9R8G+K3JPKz2OOLu1O0FZcXi25ocDFY7zmMgeKxBsDG0C77zq6Ja4sZr
+V+GI7ReXV2HmODOW92SnIC0l5vk+Fjv+z6wTWDHJqcRTlyyjGBStcz2Pz/bq0/cF
+Ueh94jaNIxnEUnahZQkElg6BIYSjKVwq3JhIWLVDzsfySp9BhDPwQiZ2fYkWWccj
+NRPf0Na3HD2I6lKBZv+lqGDSwuyofI0x8gVolq0tHqIZlxSfg3LAKBGzO2efiXc2
+yugOaaLel6lrTQqZBLR4n4qF7MOD+PmEcx2N7q0rsw==
+-----END CERTIFICATE-----
