+++ Certificates/server-key.pem
@@ -1,0 +1,5 @@
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIMWwJHJ9wiw7HL0+Bmn4AiJqKIDNhnL3yN/l0+iT2f5GoAoGCCqGSM49
+AwEHoUQDQgAEmQJOauFtWbG/134EO830rkl0DlbUFVgnF6vFvZKJ0XEPPsRYXHC7
+U0DO8GrsOvgLBEpK5EkhjZBN1Z/LuMF5xQ==
+-----END EC PRIVATE KEY-----
