+++ Certificates/server.pem
@@ -1,0 +1,21 @@
+-----BEGIN CERTIFICATE-----
+MIIDXjCCAkagAwIBAgIUQ657fKlJiizgXHYaT7v/cRVnxz4wDQYJKoZIhvcNAQEL
+BQAwdTELMAkGA1UEBhMCVVMxFjAUBgNVBAgTDVNhbiBGcmFuY2lzY28xCzAJBgNV
+BAcTAkNBMRgwFgYDVQQKEw9NeSBDb21wYW55IE5hbWUxEzARBgNVBAsTCk9yZyBV
+bml0IDIxEjAQBgNVBAMTCU15IG93biBDQTAeFw0yMjA1MTgwMDIwMDBaFw0yNzA1
+MTcwMDIwMDBaMEYxCzAJBgNVBAYTAlVTMQswCQYDVQQIEwJDQTEWMBQGA1UEBxMN
+U2FuIEZyYW5jaXNjbzESMBAGA1UEAxMJbG9jYWxob3N0MFkwEwYHKoZIzj0CAQYI
+KoZIzj0DAQcDQgAEmQJOauFtWbG/134EO830rkl0DlbUFVgnF6vFvZKJ0XEPPsRY
+XHC7U0DO8GrsOvgLBEpK5EkhjZBN1Z/LuMF5xaOB3zCB3DAOBgNVHQ8BAf8EBAMC
+BaAwHQYDVR0lBBYwFAYIKwYBBQUHAwEGCCsGAQUFBwMCMAwGA1UdEwEB/wQCMAAw
+HQYDVR0OBBYEFAOZeq64HZtaKCGnHhzXR+CHKynbMB8GA1UdIwQYMBaAFEmxalBv
+GqGYiVGimextD+g6Dtt2MF0GA1UdEQRWMFSCC2V4YW1wbGUubmV0gg93d3cuZXhh
+bXBsZS5uZXSCCWxvY2FsaG9zdIIIZnJvbnRlbmSCB2hpc3RvcnmCBndvcmtlcoII
+bWF0Y2hpbmeHBH8AAAEwDQYJKoZIhvcNAQELBQADggEBAJGScx64DdXZJMy1Gdo5
+xj3GiDVYAIdgUnpIx9plKVAUrENziTHTxVLfz2FDAobxqaMQ7iLy1qDGXO+Kl5au
+MTfaN2MS646I0dMiEe3DnUJFiGhd+Jau5Pi7exltRaRzmcSziUttEOunsTV7U9Ed
+h55ThCu98bigSKWxOz68ueCd8qxl1Nm+ViRBZvAbGi48vmxVddK8W/Cf7VQ3C4Fm
+27brqtfQ6bUEb4NMBlZdzvnGxAlUOYjhQErqnALX3f7aNYtC9ZfMSUV1g+bxIwmz
+lSLbp5TTVvVb/DWBgRFK4/AyFY4xpUqOcZRVuGECACwXYk/lvVXDVHMJWYqrWecv
+KIU=
+-----END CERTIFICATE-----
